Fixed binary message serialization
 * It turns out that the issue was coming from `EntityContainer` using `usize::MAX` as a marker for the entity container stuff.  I looked at the error the `bincode` was throwing and found out it was the max 64-bit unsigned integer, and after a while the vague memory of me using that for the entity container while sitting in Uptown Cafe at that table in the corner came to my mind and sure enough, there it was.
 * Worked around a bug that happens when Rust compacts `[u8; 3]` into a single value an Emscripten doesn't know what to do about it by changing `HybridClient`'s `color_calculator` to return `[u8; 4]` instead of `Color`
 * Everything works as expected.  Snapshots are fetched and displayed properly, and I can now ~~rest happily~~ start on the literal hours of work I've been putting off while manically working on this.

+++ minutiae/src/container.rs
@@ -3,7 +3,7 @@
 use std::collections::BTreeMap;
 use std::mem;
 use std::ops::{Index, IndexMut};
-use std::usize;
+use std::u32;
 
 #[cfg(feature = "serde")]
 use serde::Deserialize;
@@ -99,9 +99,11 @@ impl<
     I: Ord + Copy,
 > EntityContainer<C, E, M, I> {
     pub fn new() -> Self {
-        // a point of `usize::MAX` indicates that the slot is the last available one.
+        // a point of `u32::MAX` indicates that the slot is the last available one.
+        // (We use `u32` instead of `usize` so that we can handle binary serialization/deserialization on
+        // 32-bit platforms such as WebAssembly and Asm.JS)
         EntityContainer{
-            entities: vec![EntitySlot::Empty(usize::MAX)],
+            entities: vec![EntitySlot::Empty(u32::MAX as usize)],
             empty_index: 0,
             positions: EntityPositions::new()
         }
@@ -110,7 +112,7 @@ impl<
     /// Inserts an entity into the container, returning its index
     pub fn insert(&mut self, entity: Entity<C, E, M>, universe_index: I) -> usize {
         let &mut EntityContainer{ref mut entities, empty_index, ref mut positions} = self;
-        let entity_index = if empty_index != usize::MAX {
+        let entity_index = if empty_index != (u32::MAX as usize) {
             let next_empty = match entities[empty_index] {
                 EntitySlot::Empty(next_empty) => next_empty,
                 _ => unreachable!(),
